move things around/clean up
