+++ source/pervasive/set_lib.rs
@@ -10,7 +10,6 @@ use crate::set::*;
 use crate::multiset::Multiset;
 #[allow(unused_imports)]
 use crate::relations::*;
-//use crate::set::set_magic;
 
 
 verus! {
@@ -84,7 +83,7 @@ impl<A> Set<A> {
     #[via_fn]
     proof fn prove_decrease_min_unique(self, r: FnSpec(A,A) -> bool)
     {
-        set_magic::<A>();
+        lemma_set_properties::<A>();
     }
 
 
@@ -111,7 +110,7 @@ impl<A> Set<A> {
 
     #[via_fn]
     proof fn prove_decrease_max_unique(self, r: FnSpec(A,A) -> bool) {
-        set_magic::<A>();
+        lemma_set_properties::<A>();
     }
 
     // pub open spec fn to_multiset(self) -> Multiset<A> {
@@ -139,7 +138,7 @@ pub proof fn lemma_singleton_size<A>(s: Set<A>)
     ensures
         s.len() == 1
 {
-    set_magic::<A>();
+    lemma_set_properties::<A>();
     let elt = choose |elt: A| s.contains(elt);
     assert(s.remove(elt).insert(elt) =~= s);
     assert(s.remove(elt) =~= Set::empty());
@@ -180,7 +179,7 @@ pub proof fn lemma_len_union_ind<A>(s1: Set<A>, s2: Set<A>)
     decreases
         s2.len(),
 {
-    set_magic::<A>();
+    lemma_set_properties::<A>();
     if s2.len() == 0 {}
     else {
         let y = choose |y: A| s2.contains(y);
@@ -286,7 +285,7 @@ pub proof fn lemma_subset_equality<A>(x: Set<A>, y: Set<A>)
     decreases 
         x.len()
 {
-    set_magic::<A>();
+    lemma_set_properties::<A>();
     if x =~= Set::<A>::empty() {
     } else {
         let e = x.choose();
@@ -308,8 +307,8 @@ pub proof fn lemma_map_size<A,B>(x: Set<A>, y: Set<B>, f: FnSpec(A) -> B)
         x.len() == y.len(), 
     decreases x.len(),
 {
-    set_magic::<A>();
-    set_magic::<B>();
+    lemma_set_properties::<A>();
+    lemma_set_properties::<B>();
     if x.len() != 0 {
         let a = x.choose();
         lemma_map_size(x.remove(a), y.remove(f(a)), f);
@@ -405,7 +404,7 @@ pub proof fn find_unique_minimal_ensures<A>(s: Set<A>, r: FnSpec(A,A) -> bool)
     decreases
         s.len(),
 {
-    set_magic::<A>();
+    lemma_set_properties::<A>();
     if s.len() == 1 {
         let x = choose |x: A| s.contains(x);
         assert(s.finite());
@@ -477,7 +476,7 @@ pub proof fn find_unique_maximal_ensures<A>(s: Set<A>, r: FnSpec(A,A) -> bool)
     decreases
         s.len(),
 {
-    set_magic::<A>();
+    lemma_set_properties::<A>();
     if s.len() == 1 {
         let x = choose |x: A| s.contains(x);
         assert(s.remove(x) =~= Set::<A>::empty());
@@ -538,6 +537,205 @@ pub proof fn find_unique_maximal_ensures<A>(s: Set<A>, r: FnSpec(A,A) -> bool)
     }
 }
 
+// Ported from Dafny prelude
+pub proof fn lemma_set_union_again1<A>(a: Set<A>, b: Set<A>)
+    ensures
+        #[trigger] a.union(b).union(b) =~= a.union(b),
+{}
+
+// Ported from Dafny prelude
+pub proof fn lemma_set_union_again2<A>(a: Set<A>, b: Set<A>)
+    ensures
+        #[trigger] a.union(b).union(a) =~= a.union(b),
+{}
+
+// Ported from Dafny prelude
+pub proof fn lemma_set_intersect_again1<A>(a: Set<A>, b: Set<A>)
+    ensures
+        #[trigger] (a.intersect(b)).intersect(b) =~= a.intersect(b),
+{}
+
+// Ported from Dafny prelude
+pub proof fn lemma_set_intersect_again2<A>(a: Set<A>, b: Set<A>)
+    ensures
+        #[trigger] (a.intersect(b)).intersect(a) =~= a.intersect(b),
+{}
+
+// Ported from Dafny prelude
+pub proof fn lemma_set_difference2<A>(s1: Set<A>, s2: Set<A>, a: A)
+    ensures
+        s2.contains(a) ==> !s1.difference(s2).contains(a),
+{}
+
+// Ported from Dafny prelude
+pub proof fn lemma_set_disjoint<A>(a: Set<A>, b: Set<A>)
+    ensures
+        a.disjoint(b) ==> ((#[trigger](a+b)).difference(a) =~= b && (a+b).difference(b) =~= a)
+{}
+
+// Dafny encodes the second clause with a single directional, although
+// it should be fine with both directions?
+// Ported from Dafny prelude
+pub proof fn lemma_set_empty_equivalency_len<A>(s: Set<A>)
+    requires
+        s.finite()
+    ensures
+        (#[trigger] s.len() == 0 <==> s == Set::<A>::empty())
+         && (s.len() != 0 ==> exists |x: A| s.contains(x)),
+{
+    assert(s.len() == 0 ==> s =~= Set::empty()) by {
+        if s.len() == 0 {
+            assert(forall |a: A| !(#[trigger] Set::empty().contains(a)));
+            assert(Set::<A>::empty().len() == 0);
+            assert(Set::<A>::empty().len() == s.len());
+            assert((exists |a: A| s.contains(a)) || (forall |a: A| !s.contains(a)));
+            if exists |a: A| s.contains(a) {
+                let a = s.choose();
+                assert(s.remove(a).len() == s.len() -1) by {
+                    axiom_set_remove_len(s, a);
+                }
+            }
+        }
+    }
+    assert(s.len() == 0 <== s =~= Set::empty());
+}
+
+// Ported from Dafny prelude
+pub proof fn lemma_set_insert_same_len<A>(s: Set<A>, a: A)
+    requires
+        s.finite(),
+    ensures
+        s.contains(a) ==> #[trigger] s.insert(a).len() =~= s.len(),
+{}
+
+// magic lemma from spec function
+// Ported from Dafny prelude
+pub proof fn lemma_set_insert_diff_len<A>(s: Set<A>, a: A)
+    requires
+        s.finite(),
+    ensures
+        !s.contains(a) ==> #[trigger] s.insert(a).len() == s.len() + 1,
+{}
+
+// Ported from Dafny prelude
+pub proof fn lemma_set_remove_len_contains<A>(s: Set<A>, a: A)
+    requires
+        s.finite(),
+    ensures
+        (s.contains(a) ==> (#[trigger] (s.remove(a).len()) == s.len() -1))
+            && (!s.contains(a) ==> s.len() == s.remove(a).len()),
+{}
+
+// Ported from Dafny prelude
+pub proof fn lemma_set_disjoint_lens<A>(a: Set<A>, b: Set<A>)
+    requires
+        a.finite(),
+        b.finite(),
+    ensures
+        a.disjoint(b) ==> #[trigger] (a+b).len() == a.len() + b.len(),
+    decreases
+        a.len(),
+{
+    if a.len() == 0 {
+        lemma_set_empty_equivalency_len(a);
+        assert(a+b =~= b);
+    }
+    else {
+        if a.disjoint(b) {
+            let x = a.choose();
+            assert(a.remove(x) + b =~= (a+b).remove(x));
+            lemma_set_remove_len_contains(a, x);
+            lemma_set_remove_len_contains(a+b, x);
+            lemma_set_disjoint_lens(a.remove(x), b);
+        }
+    }
+
+}
+
+// Ported from Dafny prelude
+/// The length of the union between two sets added to the length of the intersection between the
+/// two sets is equal to the sum of the lengths of the two sets. 
+pub proof fn lemma_set_intersect_union_lens<A>(a: Set<A>, b: Set<A>)
+    requires
+        a.finite(),
+        b.finite(),
+    ensures
+        #[trigger] (a+b).len() + #[trigger] a.intersect(b).len() == a.len() + b.len(),
+    decreases
+        a.len(),
+{
+    if a.len() == 0 {
+        lemma_set_empty_equivalency_len(a);
+        assert(a+b =~= b);
+        assert(a.intersect(b) =~= Set::empty());
+        assert(a.intersect(b).len() == 0);
+    }
+    else {
+        let x = a.choose();
+        lemma_set_remove_len_contains(a, x);
+        lemma_set_intersect_union_lens(a.remove(x), b);
+        if (b.contains(x)) {
+            assert(a.remove(x) + b =~= (a+b));
+            assert(a.intersect(b).remove(x) =~= a.remove(x).intersect(b));
+            lemma_set_remove_len_contains(a.intersect(b), x);
+        }
+        else {
+            assert(a.remove(x) + b =~= (a+b).remove(x));
+            assert(a.remove(x).intersect(b) =~= a.intersect(b));
+        }
+    }
+}
+
+// Ported from Dafny prelude
+pub proof fn lemma_set_difference_len<A>(a: Set<A>, b: Set<A>)
+    requires
+        a.finite(),
+        b.finite(),
+    ensures
+        (#[trigger] a.difference(b).len() + b.difference(a).len() + a.intersect(b).len() == (a+b).len()) 
+            && (a.difference(b).len() == a.len() - a.intersect(b).len()),
+    decreases
+        a.len(),
+{
+    if a.len() == 0 {
+        lemma_set_empty_equivalency_len(a);
+        assert(a.difference(b) =~= Set::empty());
+        assert(b.difference(a) =~= b);
+        assert(a.intersect(b) =~= Set::empty());
+        assert(a+b =~= b);
+    }
+    else {
+        let x = a.choose();
+        lemma_set_difference_len(a.remove(x), b);
+        if b.contains(x) {
+            assert(a.intersect(b).remove(x) =~= a.remove(x).intersect(b));
+            assert(a.intersect(b).contains(x));
+            assert(!a.difference(b).contains(x));
+            assert(a.remove(x).difference(b) =~= a.difference(b));
+            assert(a.remove(x).difference(b).len() == a.difference(b).len());
+
+            assert(!b.difference(a).contains(x));
+            assert(b.difference(a.remove(x)).contains(x));
+            assert(b.difference(a.remove(x)).remove(x) =~= b.difference(a));
+            lemma_set_remove_len_contains(b.difference(a.remove(x)),x);
+            assert(b.difference(a.remove(x)).len() == b.difference(a).len() + 1);
+
+            assert(a.intersect(b).contains(x));
+            assert(!a.remove(x).intersect(b).contains(x));
+            assert(a.remove(x).intersect(b) =~= a.intersect(b).remove(x));
+            lemma_set_remove_len_contains(a.intersect(b),x);
+
+            assert(a.remove(x) + b =~= a+b);
+            assert(a.difference(b).len() + b.difference(a).len() + a.intersect(b).len() == (a+b).len());
+        }
+        else {
+            assert(!a.remove(x).difference(b).contains(x));
+            assert(a.remove(x).difference(b).len() == a.difference(b).len() -1);
+            assert(a.difference(b).len() + b.difference(a).len() + a.intersect(b).len() == (a+b).len());
+        }
+        assert(a.difference(b).len() == a.len() - a.intersect(b).len());
+    }
+}
 
 // pub proof fn lemma_multiset_from_set<A>()
 //     ensures
@@ -546,6 +744,70 @@ pub proof fn find_unique_maximal_ensures<A>(s: Set<A>, r: FnSpec(A,A) -> bool)
 //             && (s.to_multiset().count(a) == 1 <==> s.contains(a)),
 // {}
 
+// magic auto style bundle of lemmas that Dafny considers when proving properties of sets
+pub proof fn lemma_set_properties<A>()
+    ensures
+        forall |a: Set<A>, b: Set<A>| #[trigger] a.union(b).union(b) == a.union(b), //lemma_set_union_again1
+        forall |a: Set<A>, b: Set<A>| #[trigger] a.union(b).union(a) == a.union(b), //lemma_set_union_again2
+        forall |a: Set<A>, b: Set<A>| #[trigger] (a.intersect(b)).intersect(b) == a.intersect(b), //lemma_set_intersect_again1
+        forall |a: Set<A>, b: Set<A>| #[trigger] (a.intersect(b)).intersect(a) == a.intersect(b), //lemma_set_intersect_again2
+        forall |s1: Set<A>, s2: Set<A>, a: A| s2.contains(a) ==> !s1.difference(s2).contains(a), //lemma_set_difference2
+        forall |a: Set<A>, b: Set<A>| a.disjoint(b) ==> ((#[trigger](a+b)).difference(a) == b && (a+b).difference(b) == a), //lemma_set_disjoint
+        forall |s: Set<A>| ((#[trigger] s.len() == 0 && s.finite()) <==> s =~= Set::empty())
+                && (s.len() != 0 ==> exists |x: A| s.contains(x)), //lemma_set_empty_equivalency_len
+        forall |s: Set<A>, a: A| (s.contains(a) && s.finite()) ==> #[trigger] s.insert(a).len() == s.len(), //lemma_set_insert_same_len
+        forall |s: Set<A>, a: A| (s.finite() && !s.contains(a)) ==> #[trigger] s.insert(a).len() == s.len() + 1, //lemma_set_insert_diff_len 
+        forall |s: Set<A>, a: A| ((s.finite() && s.contains(a)) ==> (#[trigger] (s.remove(a).len()) == s.len() -1))
+                && (!s.contains(a) ==> s.len() == s.remove(a).len()), //lemma_set_remove_len_contains
+        forall |a: Set<A>, b: Set<A>| (a.finite() && b.finite() && a.disjoint(b)) ==> #[trigger] (a+b).len() == a.len() + b.len(), //lemma_set_disjoint_lens
+        forall |a: Set<A>, b: Set<A>| (a.finite() && b.finite()) ==> #[trigger] (a+b).len() + #[trigger] a.intersect(b).len() == a.len() + b.len(), //lemma_set_intersect_union_lens
+        forall |a: Set<A>, b: Set<A>| (a.finite() && b.finite()) ==> ((#[trigger] a.difference(b).len() + b.difference(a).len() + a.intersect(b).len() == (a+b).len()) 
+                && (a.difference(b).len() == a.len() - a.intersect(b).len())), //lemma_set_difference_len
+{
+    assert forall |a: Set<A>, b: Set<A>| #[trigger] a.union(b).union(b) == a.union(b) by {
+        lemma_set_union_again1(a, b);
+    }
+    assert forall |a: Set<A>, b: Set<A>| #[trigger] a.union(b).union(a) == a.union(b) by {
+        lemma_set_union_again2(a, b);
+    }
+    assert forall |a: Set<A>, b: Set<A>| #[trigger] (a.intersect(b)).intersect(b) == a.intersect(b) by {
+        lemma_set_intersect_again1(a, b);
+    }
+    assert forall |a: Set<A>, b: Set<A>| #[trigger] (a.intersect(b)).intersect(a) == a.intersect(b) by {
+        lemma_set_intersect_again2(a, b);
+    }
+    assert forall |a: Set<A>, b: Set<A>| a.disjoint(b) implies ((#[trigger](a+b)).difference(a) == b && (a+b).difference(b) == a) by {
+        lemma_set_disjoint(a, b);
+    }
+    assert forall |s: Set<A>| #[trigger] s.len() == 0 && s.finite() implies s =~= Set::empty() by {
+        lemma_set_empty_equivalency_len(s);
+    }
+    assert forall |s: Set<A>, a: A| (s.contains(a) && s.finite()) implies #[trigger] s.insert(a).len() == s.len() by {
+        lemma_set_insert_same_len(s, a);
+    }
+    assert forall |s: Set<A>, a: A| (s.finite() && !s.contains(a)) implies #[trigger] s.insert(a).len() == s.len() + 1 by {
+        lemma_set_insert_diff_len(s,a);
+    }
+    assert forall |s: Set<A>, a: A| s.contains(a) && s.finite() implies (#[trigger] (s.remove(a).len()) == s.len() -1) by {
+        lemma_set_remove_len_contains(s, a);
+    }
+    assert forall |a: Set<A>, b: Set<A>| a.disjoint(b) && b.finite() && a.finite() implies #[trigger] (a+b).len() == a.len() + b.len() by {
+        lemma_set_disjoint_lens(a, b);
+    }
+    assert forall |a: Set<A>, b: Set<A>| a.finite() && b.finite() implies #[trigger] (a+b).len() + #[trigger] a.intersect(b).len() == a.len() + b.len() by {
+        lemma_set_intersect_union_lens(a,b);
+    }
+    assert forall |a: Set<A>, b: Set<A>| (a.finite() && b.finite()) ==> #[trigger] a.difference(b).len() + b.difference(a).len() + a.intersect(b).len() == (a+b).len() by {
+        if a.finite() && b.finite() {
+            lemma_set_difference_len(a, b);
+        }
+    }
+    assert forall |a: Set<A>, b: Set<A>| (a.finite() && b.finite()) ==> #[trigger] a.difference(b).len() == a.len() - a.intersect(b).len() by {
+        if a.finite() && b.finite() {
+            lemma_set_difference_len(a, b);
+        }
+    }
+}
 
 #[doc(hidden)]
 #[verifier(inline)]
